[FIX] medor_delivery_address: Let module be installable
Not sure what it did here… I don't understand the utility of the line:
```
<attribute name="groups">!sale.group_delivery_invoice_address</attribute>
```
It seams not to work. I understant the `!` as a negation sign but I
can find other similar example.

I really don't like to do a replace in xml view. So for me a good way is
to modify classes for the existing address in the sexy invoice and then
add the shipping address if it is different.