[ADD] pos_receipt_xml_header_footer
In Odoo 12, there exist two XML templates that are both receipts:
PosTicket and XmlReceipt. PosTicket is shown on the cashier's screen,
and the more tightly defined XmlReceipt is sent to the ESC POC printer.
In future Odoo versions, I believe they have been unified into
OrderReceipt.

For reasons unknown to me, PosTicket does not contain support for
`header_xml` and `footer_xml`, but XmlReceipt does. This module fixes
that.

Signed-off-by: Carmen Bianca Bakker <carmen@coopiteasy.be>