[ADD] delivery_carrier_combine_price_rule
When adding price rule to a delivery carrier one can chose how price
rule will be combined. By default it's a OR combinaison, meaning it
stops after the first condition that is true.

Now one can chose AND combinaison, meaning it sums the price for all the
price rules and it cannot compute if there is one rule that is false.