[IMP] account_customer_wallet : make the module working with global partners
**before**

the computation is based on that account  ``customer_wallet_account_id =
fields.Many2one(related="company_id.customer_wallet_account_id")``

so, if you have a "global" partner, there is no company so no account,
and the wallet feature will not work.

**After**

The computation is based on ``wallet_account =
self.env.company.customer_wallet_account_id``. so it depends on the
context. It will work for global partners, (no company_id) the value
will depend on the current context. (that makes sense in my opinion). If
you are in a company A, you have a CoA A and you have a wallet A amount.
if you are in a company B, you have a coA B and you have a wallet B
amount.