[IMP] account_customer_wallet: Verify account.move is posted before including it in balance calculation
The reason this test is added (apart from being objectively more
correct) is that this work is necessary to get the account.payment stuff
right. _create_payments() in account.payment.register creates the
necessary account.move.lines _before_ we do our balance test in
action_post() in account.payment.

This (almost) completely refactored
get_wallet_balance_account_move_line() to no longer use the where_query
stuff. For some reason, when adding the extra state domain check, the
database query returned nothing. So logically, one opens the debugger
and does a manual search on the domain in that method, before the
database query is executed. This returned the expected result.
Subsequently, when continuing the debugger, the database query would
return the expected result...

I wasn't able to investigate the cause of that discrepancy, so I just
rewrote the method to do a plain search on the database. Is there a
reason we didn't do this before? I can find this discussion:
https://github.com/coopiteasy/addons/pull/222/files#r856062451 belonging
to this commit:
https://github.com/coopiteasy/addons/commit/851a2b8e12bc6f30eae4c9d75ab667e832adc291

But the only identifiable benefit is being able to query multiple
partners, which we already do.

Signed-off-by: Carmen Bianca BAKKER <carmen@coopiteasy.be>