[REF] resource_work_time_from_contracts: remove readonly
This line has the side effect of issuing a warning
for all models inheriting resource.mixin and setting defaults
to resource_calendar_id.

The readonly was introduced to prevent modifications by the user
since the modules computes it. It is however to relevant since
the field is hidden (readonly parameter only affects the views).