[FIX] website_sale_product_contract_gift: "copy" in name
When using module "partner_firstname", the copy() method adds a
"(copy)" string to the firstname. Even if writing the name when copying.

Also when copying all fields are copied, and in this case we want only
the fields that are filled in the form on the website.

So using create() is better than copy() for this case.